% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set_praat_location}
\alias{set_praat_location}
\alias{get_praat_location}
\title{Get or set the location of the Praat executable}
\usage{
set_praat_location(path)

get_praat_location()
}
\arguments{
\item{path}{the full path to the Praat executable}
}
\value{
the full path to the Praat executable
}
\description{
Get or set the location of the Praat executable
}
\details{
\code{set_praat_location()} sets an R \code{\link[=options]{options()}} for
\code{options("tjm.praat_location")}. \code{get_praat_location()} retrieves the value
of this option.
\subsection{Default location}{

On package load, this package checks for a pre-existing value in
\code{options("tjm.praat_location")} and defers to that value. If there is no such value,
this package then searches the system PATH using \code{\link[=Sys.which]{Sys.which()}} on \code{"praat"}.
If Praat is not found on the PATH, \code{NULL} is used.

A user can set a default location for Praat by 1) adding it to their PATH.
For example, I use a folder called \code{bin} in my Documents folder. My copy of
Praat.exe lives in this folder, and this folder is on my user PATH on
Windows. Alternatively, a user can 2) add a line like
\verb{options(tjm.praat_location = [value])} to a .Rprofile file so that the
option is set at the start of each R session.
}
}
