% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_spectrogram}
\alias{read_spectrogram}
\title{Read a Praat spectrogram file produced by Praat}
\usage{
read_spectrogram(path)
}
\arguments{
\item{path}{path to a .Spectrogram file}
}
\value{
a dataframe with the Spectrogram data
}
\description{
Read a Praat spectrogram file produced by Praat
}
