% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set_textgrid_ext}
\alias{set_textgrid_ext}
\title{Set the file-extension in a path to \code{.TextGrid}}
\usage{
set_textgrid_ext(xs)
}
\arguments{
\item{xs}{paths or filenames}
}
\value{
the paths with their extensions replaced with \code{.TextGrid}
}
\description{
Set the file-extension in a path to \code{.TextGrid}
}
\examples{
set_textgrid_ext("C:/Programs/Nullsoft/Winamp/Demo.mp3")
}
