% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{wrap_praat_script}
\alias{wrap_praat_script}
\title{Make a function that runs a Praat script}
\usage{
wrap_praat_script(
  praat_location,
  script_code_to_run,
  return = c("last-argument", "info-window")
)
}
\description{
Make a function that runs a Praat script
}
