% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{create_silences_textgrid}
\alias{create_silences_textgrid}
\title{Create a silences textgrid}
\format{
A Praat script
\describe{
\item{wav_file_in}{path for the wave to read}
\item{textGrid_out}{path of the textgrid file to create}
}
}
\usage{
create_silences_textgrid
}
\description{
Create a silences textgrid
}
\section{Example usage}{
\if{html}{\out{<div class="r">}}\preformatted{library(tjm.praat)
library(purrr)
library(readtextgrid)

# Use the PATH to find praat
praat_path <- Sys.which("praat")

# Locate some wav files

dir <- system.file("demo-textgrids/", package = "tjm.praat")
wavs <- list.files(dir, pattern = ".wav", full.names = TRUE)

# Put the resulting textgrids into a temporary folder
tgs <- file.path(tempdir(), basename(set_textgrid_ext(wavs)))

f_create_silences_textgrid <- wrap_praat_script(
  praat_location = praat_path,
  script_code_to_run = create_silences_textgrid,
  return = "last-argument"
)

# Apply the function to each wav file and read into R
map2_chr(wavs, tgs, f_create_silences_textgrid) \%>\%
  map_df(read_textgrid)
#> # A tibble: 0 x 0
}\if{html}{\out{</div>}}
}

\section{Source code}{


#' ```{r, echo = FALSE, comment = ""}
create_silences_textgrid\preformatted{}
}

\keyword{datasets}
