% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\docType{data}
\name{create_silences_textgrid}
\alias{create_silences_textgrid}
\title{Create a silences textgrid}
\format{
A Praat script
\describe{
\item{wav_file_in}{path for the wave to read}
\item{textgrid_out}{path of the textgrid file to create}
}
}
\usage{
create_silences_textgrid
}
\description{
Create a silences textgrid
}
\section{Example usage}{
\if{html}{\out{<div class="r">}}\preformatted{library(tjm.praat)
library(purrr)
library(readtextgrid)

# Use the PATH to find praat
praat_path <- Sys.which("praat")

# Locate some wav files

dir <- system.file("test-wavs-a/", package = "tjm.praat")
wavs <- list.files(dir, pattern = ".wav", full.names = TRUE)

# Put the resulting textgrids into a temporary folder
tgs <- file.path(tempdir(), basename(set_textgrid_ext(wavs)))

f_create_silences_textgrid <- wrap_praat_script(
  praat_location = praat_path,
  script_code_to_run = create_silences_textgrid,
  return = "last-argument"
)

# Apply the function to each wav file and read into R
map2_chr(wavs, tgs, f_create_silences_textgrid) \%>\%
  map_df(read_textgrid)
#> # A tibble: 6 x 10
#>   file  tier_num tier_name tier_type tier_xmin tier_xmax  xmin  xmax text 
#>   <chr>    <dbl> <chr>     <chr>         <dbl>     <dbl> <dbl> <dbl> <chr>
#> 1 the_~        1 silences  Interval~         0       0.6 0     0.464 soun~
#> 2 the_~        1 silences  Interval~         0       0.6 0.464 0.6   sile~
#> 3 the_~        1 silences  Interval~         0       0.6 0     0.464 soun~
#> 4 the_~        1 silences  Interval~         0       0.6 0.464 0.6   sile~
#> 5 the_~        1 silences  Interval~         0       0.6 0     0.464 soun~
#> 6 the_~        1 silences  Interval~         0       0.6 0.464 0.6   sile~
#> # ... with 1 more variable: annotation_num <int>
}\if{html}{\out{</div>}}
}

\section{Praat source code}{
\preformatted{form Make silence text grid
    sentence Wav_file_in
    sentence Textgrid_out
endform
Read from file: wav_file_in$
To TextGrid (silences): 100, 0, -25, 0.1, 0.1, "silent", "sounding"
Save as text file: textgrid_out$
}
}

\keyword{datasets}
