% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{wrap_praat_script}
\alias{wrap_praat_script}
\title{Make a function that runs a Praat script}
\usage{
wrap_praat_script(
  script_code_to_run,
  returning = c("last-argument", "info-window"),
  praat_location = get_praat_location()
)
}
\arguments{
\item{script_code_to_run}{a Praat script to run.}

\item{returning}{value to return. \code{"last-argument"} returns the last argument to
the Praat script. \code{"info-window"} returns the contents of the Praat Info
Window.}

\item{praat_location}{path to the Praat executable. Defaults to the value
provided by \code{\link[=get_praat_location]{get_praat_location()}}.}
}
\value{
see \code{return} argument
}
\description{
Make a function that runs a Praat script
}
\details{
This function basically sets up a call to Praat's command-line
interface using \code{system2()}.
}
